Made the unit value, line value and total value elements optional in a credit request document.
This is to cater for credit request lines where there is just an incorrect number of boxes on a supplier boxes line in a consolidated invoice. Supplier "boxes" lines are special lines on a consolidated invoice that represent an additional supplier. The value of the credit request is not known in this scenario.

bugid: 595