Branched from tsMappingHospitalityCocaColaPurchaseOrder.xsl in order to fix certain values that would otherwise change as part of case 3272 (FnB trading relationship re-organisation)
Bighams have failed to engage with Fourth on what impact if any the 3272 changes will have on their system. As the migration for 3272 can't wait any longer, the current values are now hard coded.

If in any future project Bighams want these values to change, please consider the impact on orders to Bighams from FnB

bugid: 3586